# public/api-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FreePic API",
    "description": "FreePic copyright-free image sharing platform API for AI agents and developers",
    "version": "1.0.0",
    "contact": {
      "name": "FreePic Dev Team",
      "email": "dev@freepic.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Development server"
    },
    {
      "url": "https://api.freepic.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/api/images": {
      "get": {
        "summary": "List images",
        "description": "Get a paginated list of approved images with optional filters",
        "tags": ["Images"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20 },
            "description": "Items per page (max 100)"
          },
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search query (searches in camera info, description)"
          },
          {
            "name": "camera",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by camera model"
          },
          {
            "name": "userId",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of images",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "images": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Image" }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/images/{id}": {
      "get": {
        "summary": "Get image details",
        "description": "Get detailed information about a specific image",
        "tags": ["Images"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Image UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Image details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Image" }
              }
            }
          },
          "404": {
            "description": "Image not found"
          }
        }
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "summary": "Get user profile",
        "description": "Get public profile information for a user",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "User UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Image": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Image UUID" },
          "url": { "type": "string", "description": "Original image URL" },
          "thumbnailUrl": { "type": "string", "description": "400px thumbnail URL" },
          "smallUrl": { "type": "string", "description": "400px small version URL" },
          "regularUrl": { "type": "string", "description": "1080px regular version URL" },
          "fullUrl": { "type": "string", "description": "1920px full version URL" },
          "width": { "type": "integer", "description": "Image width in pixels" },
          "height": { "type": "integer", "description": "Image height in pixels" },
          "author": { "type": "string", "description": "Display name of photographer" },
          "userId": { "type": "string", "description": "UUID of the photographer" },
          "camera": { "type": "string", "description": "Camera model used" },
          "likes": { "type": "integer" },
          "downloads": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "blurHash": { "type": "string", "description": "Blur placeholder hash" },
          "dominantColor": { "type": "string", "description": "Hex color code" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "avatar": { "type": "string" },
          "bio": { "type": "string" },
          "location": { "type": "string" },
          "website": { "type": "string" },
          "instagram": { "type": "string" },
          "twitter": { "type": "string" }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      }
    }
  }
}